requirements.txt
flask<=2
psycopg2-binary<=3
python-dotenv==0.18.0
